Default to 0 for concentrations
Having a default value other than 0 results in a post to the api without
said component will default to other than 0. This is very confusing.

The reason for having default other than 0 is likely to show some
numbers when used. In formconfig however we do show other values, let's
leave it at that.