Add missing substances in Concentrations
The main issue here is that we return an instance of the pydantic
Concentrations, and if the substance is not included in the pydantic
object, it will not be included in the response either.

Tecnically the entries in Concentrations should be one to one with
entries in "reactions.Molecule". For now added a test to ensure this